+++ src/Settings/AppearancePage.cpp
@@ -19,13 +19,8 @@
 #include "Core/Translator.hpp"
 #include "Settings/SettingsManager.hpp"
 #include "Settings/ValueWrapper.hpp"
-#include <QApplication>
 #include <QMessageBox>
 #include <generated/SettingsInfo.hpp>
-
-#ifdef Q_OS_WIN
-#include "Util/Util.hpp"
-#endif
 
 AppearancePage::AppearancePage(QWidget *parent)
     : PreferencesPageTemplate({"Locale", "Editor Theme", "Editor Font", "Test Cases Font", "Message Logger Font",
@@ -46,16 +41,6 @@ void AppearancePage::makeSettingsTheSameAsUI()
             QMessageBox::warning(this, tr("Change Locale"),
                                  tr("You need to restart the application to completely apply the locale change."));
         }
-#ifdef Q_OS_WIN
-        if (si.name == "Editor Theme" && SettingsManager::get(si.name) != widget->getVariant() &&
-            Util::windowsDarkThemePalette() == qApp->palette())
-        {
-            QMessageBox::warning(
-                this, tr("Change Editor Theme"),
-                tr("Because you are using whole-application dark theme on Windows, you need to restart "
-                   "the application after changing the editor theme."));
-        }
-#endif
         SettingsManager::set(si.name, widget->getVariant());
     }
 }