+++ doc/CHANGELOG.md
@@ -26,12 +26,14 @@
 - Fix a bug which makes the application crash when quiting in some scenarios. (fa5259b)
 - Fix a bug which makes the saved test cases are not loaded correctly.
 - Fix the screen is not scrolled to show the cursor in some situations.
+- Fix change of Editor theme on Windows Dark mode required a restart of application. (#232)
 
 ### Changed
 
 - Open an empty untitled tab when the open file length limit is exceeded. (#353)
 - When parsing a problem while it is already parsed and opened in a tab, the new test cases will override the old test cases.
 - If you click "Cancel" when saving all files, the action will be aborted instead of saving the next file.
+- The Qt version is upgraded to 5.15. (#374 and #378)
 
 ## v6.4